+++ webapp/views/App/views/Designer/Taxonomy/EditHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './EditHeader'
