+++ webapp/views/App/views/Designer/Taxonomy/TaxaTableRow/TaxaTableRow.js
@@ -1,10 +1,11 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import * as R from 'ramda'
 
 import * as Taxon from '@core/survey/taxon'
 import * as TaxonVernacularName from '@core/survey/taxonVernacularName'
 
-const TaxaTableRow = props => {
+const TaxaTableRow = (props) => {
   const { row: taxon, idx, offset, vernacularLanguageCodes } = props
 
   return (
@@ -14,7 +15,7 @@ const TaxaTableRow = props => {
       <div>{Taxon.getFamily(taxon)}</div>
       <div>{Taxon.getGenus(taxon)}</div>
       <div>{Taxon.getScientificName(taxon)}</div>
-      {vernacularLanguageCodes.map(lang => {
+      {vernacularLanguageCodes.map((lang) => {
         const vernacularNames = Taxon.getVernacularNamesByLang(lang)(taxon)
         return (
           <div key={`${Taxon.getUuid(taxon)}_vernacular_name_${lang}`}>
@@ -24,6 +25,12 @@ const TaxaTableRow = props => {
       })}
     </>
   )
+}
+TaxaTableRow.propTypes = {
+  row: PropTypes.object.isRequired,
+  idx: PropTypes.number.isRequired,
+  offset: PropTypes.number.isRequired,
+  vernacularLanguageCodes: PropTypes.array.isRequired,
 }
 
 export default TaxaTableRow
