+++ webapp/views/App/views/Designer/Taxonomy/TaxaTableRow/index.js
@@ -1,0 +1,1 @@
+export { default } from './TaxaTableRow'
