+++ webapp/views/App/views/Designer/Taxonomy/TaxaTableRowHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './TaxaTableRowHeader'
