+++ webapp/views/App/views/Designer/Taxonomy/index.js
@@ -1,0 +1,1 @@
+export { default } from './Taxonomy'
