+++ webapp/views/App/views/Designer/Taxonomy/reducer.js
@@ -2,7 +2,6 @@ import { exportReducer } from '@webapp/utils/reduxUtils'
 
 import { SystemActions } from '@webapp/store/system'
 import { SurveyActions } from '@webapp/store/survey'
-import { formReset } from '../surveyForm/actions'
 
 import { taxonomyViewTaxonomyPropsUpdate, taxonomyViewTaxonomyUpdate } from './actions'
 import * as TaxonomyState from './taxonomyState'
@@ -14,7 +13,6 @@ const actionHandlers = {
   [SurveyActions.surveyCreate]: () => ({}),
   [SurveyActions.surveyUpdate]: () => ({}),
   [SurveyActions.surveyDelete]: () => ({}),
-  [formReset]: () => ({}),
 
   [taxonomyViewTaxonomyUpdate]: (_state, { taxonomyUuid }) => TaxonomyState.initTaxonomyEdit(taxonomyUuid),
