+++ webapp/views/App/views/Designer/TaxonomyList/HeaderLeft/HeaderLeft.js
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux'
 import { appModuleUri, designerModules } from '@webapp/app/appModules'
 
 import { matchPath, useHistory, useLocation } from 'react-router'
-import * as TaxonomyActions from '@webapp/loggedin/surveyViews/taxonomy/actions'
+import * as TaxonomyActions from '@webapp/views/App/views/Designer/Taxonomy/actions'
 
 import * as Taxonomy from '@core/survey/taxonomy'
