+++ webapp/views/App/views/Designer/TaxonomyList/Row/store/actions/useEdit.js
@@ -5,7 +5,7 @@ import { matchPath, useHistory, useLocation } from 'react-router'
 import * as ObjectUtils from '@core/objectUtils'
 import { appModuleUri, designerModules } from '@webapp/app/appModules'
 
-import * as TaxonomyActions from '@webapp/loggedin/surveyViews/taxonomy/actions'
+import * as TaxonomyActions from '@webapp/views/App/views/Designer/Taxonomy/actions'
 import * as Taxonomy from '@core/survey/taxonomy'
 
 import { State } from '../state'
